requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
channels==4.0.0
Django==4.2.7
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
